v0.14.0
ReflectPath.find_modules will now ignore found modules that don't actually exist (meaning the name is correct but it's not actually a python module).

CallbackServer now lets the code be set and the error handling is a little more general.

Cleans up comments in the string module.

Fixes a bug in Url.parts where it returned an empty string if there wasn't a path